Read data from YAML and it actually works!

+++ app/models/Shot.java
@@ -2,9 +2,9 @@ package models;
 
 public class Shot {
 	public int no;
-	public long start, end;
+	public int start, end;
 
-	public Shot(int no, long start, long end) {
+	public Shot(int no, int start, int end) {
 		this.no = no;
 		this.start = start;
 		this.end = end;
